Modify : VulDBPin, VulDBShare 컴포넌트 추출
