+++ src/components/vulnerability-db/ArticleDetailHeader.tsx
@@ -5,9 +5,9 @@ import { useVulDBPostsStore } from "@/stores/useVulDBPostsStore";
 import { VulDBPinnedInfo, VulDBPost } from "@/types/post";
 import { Timestamp } from "firebase/firestore";
 import { useLayoutEffect, useMemo, useState } from "react";
-import { IconShare } from "../ui/Icons";
 import { Label } from "../ui/Label";
 import VulDBPin from "./VulDBPin";
+import VulDBShare from "./VulDBShare";
 
 export function LabelSkeleton() {
   return (
@@ -25,7 +25,6 @@ export function ArticleDetailHeader({
   const haveUpdatedAt = post.source_updated_at;
   const { vulDBPostsWithChip } = useVulDBPostsStore();
   const [isMounted, setIsMounted] = useState(false);
-  const [isCopied, setIsCopied] = useState(false);
 
   useLayoutEffect(() => {
     setIsMounted(true);
@@ -37,17 +36,6 @@ export function ArticleDetailHeader({
   }, [vulDBPostsWithChip, post.id]);
 
   const pinnedInfo: VulDBPinnedInfo = { userId, postId: post.id };
-
-  const handleShareClick = async () => {
-    try {
-      const currentUrl = window.location.href;
-      await navigator.clipboard.writeText(currentUrl);
-      setIsCopied(true);
-      setTimeout(() => setIsCopied(false), 2000); // 2초 후에 복사 상태를 초기화
-    } catch (error) {
-      console.error("URL 복사 실패:", error);
-    }
-  };
 
   return (
     <section className="w-full max-w-[82.125rem] border-b border-b-line-default p-[1.75rem_0_3.75rem_0] px-[1rem]">
@@ -77,15 +65,7 @@ export function ArticleDetailHeader({
         </div>
         <div className="relative flex gap-[1.625rem]">
           <VulDBPin pinnedInfo={pinnedInfo} />
-          <IconShare onClick={handleShareClick} />
-          {isCopied && (
-            <div
-              className="absolute top-full mt-2 w-[210px] rounded-lg bg-white p-2 text-center text-[20px] text-green-500 shadow-lg"
-              style={{ left: "-25px" }}
-            >
-              URL이 복사되었습니다.
-            </div>
-          )}
+          <VulDBShare postId={post.id} />
         </div>
       </div>
     </section>
