+++ src/components/vulnerability-db/SimilarInfoPosts.tsx
@@ -7,7 +7,6 @@ import {
   CardSubTitle,
   CardTitle,
 } from "@/components/ui/Card";
-import { IconShare } from "@/components/ui/Icons";
 import { Label } from "@/components/ui/Label";
 import { formatTimestampAsDaysAgo } from "@/lib/utils";
 import { useVulDBPostsStore } from "@/stores/useVulDBPostsStore";
@@ -17,6 +16,7 @@ import Link from "next/link";
 import { useLayoutEffect, useMemo, useState } from "react";
 import { LabelSkeleton } from "./ArticleDetailHeader";
 import VulDBPin from "./VulDBPin";
+import VulDBShare from "./VulDBShare";
 
 function SkeletonCard() {
   return (
@@ -53,20 +53,6 @@ function SimilarInfoPostCard({
 
   const pinnedInfo: VulDBPinnedInfo = { userId, postId: post.id };
 
-  // 게시물에 대한 URL 생성 (예: /posts/[postId])
-  const postUrl = `${window.location.origin}/vuldb/items/${post.id}`;
-
-  const handleShareClick = async (event: React.MouseEvent) => {
-    event.stopPropagation();
-    event.preventDefault();
-    try {
-      await navigator.clipboard.writeText(postUrl);
-      alert("URL이 복사되었습니다!"); // 복사가 완료되면 알림 표시
-    } catch (error) {
-      console.error("URL 복사에 실패했습니다.", error);
-    }
-  };
-
   return (
     <Card
       variant="article"
@@ -99,8 +85,8 @@ function SimilarInfoPostCard({
       {/* CNNVD 데이터가 db에 들어오면 추가 예정 */}
       <CardFooter>
         <div className="inline-flex gap-4">
-         <VulDBPin pinnedInfo={pinnedInfo} />
-          <IconShare onClick={handleShareClick} />
+          <VulDBPin pinnedInfo={pinnedInfo} />
+          <VulDBShare postId={post.id} />
         </div>
         <CardSubTitle className="font-medium">
           {post.created_at
