+++ src/components/vulnerability-db/VulDBList.tsx
@@ -6,16 +6,15 @@ import {
   CardSubTitle,
   CardTitle,
 } from "@/components/ui/Card";
-import { IconExternalLink } from "@/components/ui/Icons";
 import { Label } from "@/components/ui/Label";
 import { increasePostViews } from "@/lib/api/posts";
-import { addPinnedPostToUser } from "@/lib/api/users";
 import { formatTimestampAsDaysAgo } from "@/lib/utils";
 import { VulDBPinnedInfo, VulDBPostWithChip } from "@/types/post";
 import { isCertCCContentType, isCnnvdContentType } from "@/types/typeGuards";
 import Link from "next/link";
 import { LabelSkeleton } from "./ArticleDetailHeader";
 import VulDBPin from "./VulDBPin";
+import VulDBShare from "./VulDBShare";
 
 export function VulDBListCardSkeleton() {
   return (
@@ -114,9 +113,9 @@ function VulDBListCard({
         </div>
       </CardContent>
       <CardFooter>
-        <div className="flex gap-3">
+        <div className="flex gap-4">
           <VulDBPin pinnedInfo={pinnedInfo} />
-          <IconExternalLink />
+          <VulDBShare postId={post.id} />
         </div>
         <CardSubTitle color="#A2A2A2" className="font-medium">
           {daysAgo || "날짜를 알 수 없음"}
