+++ src/components/vulnerability-db/VulDBPin.tsx
@@ -1,27 +1,34 @@
+"use client";
+
 import { addPinnedPostToUser } from "@/lib/api/users";
 import { VulDBPinnedInfo } from "@/types/post";
-import { IconPin } from "../ui/Icons";
+import { IconFilledPin, IconPin } from "../ui/Icons";
+import { useState } from "react";
 
 export default function VulDBPin({
   pinnedInfo,
 }: {
   pinnedInfo: VulDBPinnedInfo;
 }) {
+  const [isClicked, setIsClicked] = useState(false);
+
   const onClickVulDBPin = async (
     event: React.MouseEvent<HTMLButtonElement>,
   ) => {
     try {
       event.stopPropagation();
       event.preventDefault();
+      setIsClicked(true);
       await addPinnedPostToUser(pinnedInfo);
     } catch (error) {
+      setIsClicked(false);
       console.error(error);
     }
   };
 
   return (
     <button onClick={onClickVulDBPin}>
-      <IconPin />
+      {isClicked ? <IconFilledPin /> : <IconPin />}
     </button>
   );
 }
