+++ src/components/vulnerability-db/VulDBShare.tsx
@@ -1,0 +1,37 @@
+"use client";
+
+import { IconFilledShare, IconShare } from "@/components/ui/Icons";
+import { useState } from "react";
+
+export default function VulDBShare({ postId }: { postId: string }) {
+  const [isCopied, setIsCopied] = useState(false);
+
+  // 게시물에 대한 URL 생성 (예: /posts/[postId])
+  const postUrl = `${window.location.origin}/vuldb/items/${postId}`;
+
+  const handleShareClick = async (event: React.MouseEvent) => {
+    event.stopPropagation();
+    event.preventDefault();
+    try {
+      await navigator.clipboard.writeText(postUrl);
+      setIsCopied(true);
+      setTimeout(() => setIsCopied(false), 2000); // 2초 후에 복사 상태를 초기화
+    } catch (error) {
+      console.error("URL 복사 실패:", error);
+    }
+  };
+
+  return (
+    <button onClick={handleShareClick}>
+      {!isCopied ? <IconShare /> : <IconFilledShare />}
+      {isCopied && (
+        <div
+          className="absolute top-full mt-2 w-[210px] rounded-lg bg-white p-2 text-center text-[20px] text-green-500 shadow-lg"
+          style={{ left: "-25px" }}
+        >
+          URL이 복사되었습니다.
+        </div>
+      )}
+    </button>
+  );
+}
